+++ lib/closure_tree/configuration.rb
@@ -1,0 +1,9 @@
+module ClosureTree
+  class Configuration # :nodoc:
+    attr_accessor :database_less
+
+    def initialize
+      @database_less = false
+    end
+  end
+end